remove mlflow dep from aim/aim2 controllers -- moved to pkg/common/api
