+++ pkg/api/admin/service/namespace/validator_test.go
@@ -6,7 +6,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"github.com/G-Research/fasttrackml/pkg/api/mlflow/api"
+	"github.com/G-Research/fasttrackml/pkg/common/api"
 )
 
 func TestValidateUpdateRunRequest_Ok(t *testing.T) {
