+++ pkg/api/mlflow/service/experiment/validators.go
@@ -1,8 +1,8 @@
 package experiment
 
 import (
-	"github.com/G-Research/fasttrackml/pkg/api/mlflow/api"
 	"github.com/G-Research/fasttrackml/pkg/api/mlflow/api/request"
+	"github.com/G-Research/fasttrackml/pkg/common/api"
 )
 
 const (
