+++ pkg/api/mlflow/service/metric/validators.go
@@ -1,8 +1,8 @@
 package metric
 
 import (
-	"github.com/G-Research/fasttrackml/pkg/api/mlflow/api"
 	"github.com/G-Research/fasttrackml/pkg/api/mlflow/api/request"
+	"github.com/G-Research/fasttrackml/pkg/common/api"
 )
 
 const (
