+++ pkg/api/mlflow/service/run/validators.go
@@ -1,8 +1,8 @@
 package run
 
 import (
-	"github.com/G-Research/fasttrackml/pkg/api/mlflow/api"
 	"github.com/G-Research/fasttrackml/pkg/api/mlflow/api/request"
+	"github.com/G-Research/fasttrackml/pkg/common/api"
 )
 
 const (
